+++ xbmc/dialogs/GUIDialogContextMenu.h
@@ -107,7 +107,7 @@ enum CONTEXT_BUTTON { CONTEXT_BUTTON_CANCELLED = 0,
                       CONTEXT_BUTTON_ACTIVE_ADSP_SETTINGS,
                     };
 
-class CContextButtons : public std::vector< std::pair<unsigned int, std::string> >
+class CContextButtons : public std::vector< std::pair<size_t, std::string> >
 {
 public:
   void Add(unsigned int, const std::string &label);
